Add files via upload
I added the "My work" area below the qualifications and interests section. I have the "my work" section very bare-bones with the links working, but the images and animations for the section can be better. Also, buttons on the top of the page still do not work properly, finding a very good solution soon.