Final Commit pt.1
I've surely added enough, the simple but well design portfolio is done. Extra content sits in the "img" folder and classes in the CSS files that aren't you use. Still want to keep them around in case of a sudden change that comes to mind, but that is all that needs to be done. This is the final product, please give me a 100 on this assignment.